# build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IResponse": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
			"defaultFisrtName": {
				"type": "string",
				"example": "firstName",
				"description": "Field for name"
			},
			"defaultLastName": {
				"type": "string",
				"example": "lastName",
				"description": "Field for lastName"
			},
			"defaultEmail": {
				"type": "string",
				"example": "kadjibecker@gmail.com",
				"description": "Field for name"
			},
			"UserType.userCreateFields": {
				"properties": {
					"firstName": {
						"$ref": "#/components/schemas/defaultFisrtName"
					},
					"lastName": {
						"$ref": "#/components/schemas/defaultLastName"
					},
					"email": {
						"$ref": "#/components/schemas/defaultEmail"
					}
				},
				"required": [
					"firstName",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "nodejsexpress",
		"version": "1.0.0",
		"description": "Nodejs-typescript backend kernel application",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/post": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Post Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"title": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"image": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"title",
									"description",
									"image"
								]
							}
						}
					}
				}
			}
		},
		"/user": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserType.userCreateFields"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}